renamed views, switching now functional
