+++ src/modules/search/subcomponents/index.ts
@@ -1,4 +1,4 @@
-export * from './grid_view';
-export * from './list_view';
+export * from './search_grid_view';
+export * from './search_list_view';
 export * from './search_navbar';
 export * from './search_sidebar';
